make Python3 compatible and add super class.
