+++ analytics_platform/kronos/uranus/src/uranus_constants.py
@@ -3,5 +3,5 @@ COMPANION_COUNT_THRESHOLD = 3
 ALTERNATE_COUNT_THRESHOLD = 2
 OUTLIER_PROBABILITY_THRESHOLD = 0.88
 UNKNOWN_PROBABILITY_THRESHOLD = 0.3
-NUM_PARTITIONS = 2
-MIN_SUPPORT_COUNT = 0.01
+NUM_PARTITIONS = 12
+MIN_SUPPORT_COUNT = 0.021
